split Pilots part from WMSAdministrator
